RPD_PATH 14_2_2026

+++ RPD_API/DTO/GrowthRate/GrowthRateDTO.cs
@@ -4,6 +4,6 @@
     {
         public Guid growthRateID { get; set; }
         public string grName { get; set; }
-        public double grTotalExp { get; set; }
+        public int grTotalExp { get; set; }
     }
 }
